+++ src/command/server/command/setup.rs
@@ -233,6 +233,7 @@ pub async fn build_registry(
         validate_manifest_references: !config.global.allow_missing_manifest_references,
         global_immutable_tags: config.global.immutable_tags,
         global_immutable_tags_exclusions: config.global.immutable_tags_exclusions.clone(),
+        listing_read_concurrency: config.global.listing_read_concurrency,
         event_dispatcher,
         ..RegistryConfig::new(job_store)
     };